% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compat.R
\docType{data}
\name{compat}
\alias{compat}
\alias{dplyr_compat}
\alias{plyr_compat}
\alias{plyr_warn_compat}
\title{Compatibility layers}
\format{\preformatted{List of 3
 $ mutate:function (.data, ...)  
 $ count :function (df, vars = NULL, wt_var = NULL)  
 $ rename:function (x, replace, warn_missing = TRUE, warn_duplicated = TRUE)  
}}
\usage{
plyr_compat

plyr_warn_compat

dplyr_compat
}
\description{
Attach these layers via \code{attach} for various levels of
compatibility to the \code{plyr} package when both \code{plyr}
and \code{dplyr} are loaded.
}
\details{
\describe{
  \item{plyr_compat}{Full compatibility, the original \code{plyr}
    implementation is used.
    You should be able to run your existing code without modifications.}
  \item{plyr_warn_compat}{Same as \code{plyr_compat},
    but emit a warning for each call of the affected functions.}
  \item{dplyr_compat}{Full compatibility with \code{dplyr}}
}
}
\keyword{datasets}

