% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap.R
\name{count}
\alias{count}
\title{Wrapped count function}
\usage{
count(df, vars = NULL, wt_var = NULL)
}
\arguments{
\item{df}{data frame to be processed}

\item{vars}{variables to count unique values of}

\item{wt_var}{optional variable to weight by - if this is non-NULL, count
will sum up the value of this variable for each combination of id
variables.}
}
\description{
This active binding returns \code{get_count_wrapper()}.
Thus, the interface and the results of calling \code{count()}
from this package depend on the currently active compatibility layer.
When the package, is loaded the \code{\link{plyr_compat}} layer is active;
this documentation reflects the interface presented by that layer.
}
\seealso{
\code{\link{pdlyr_compat}}
}

