% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap.R
\name{forward}
\alias{dcount}
\alias{dmutate}
\alias{drename}
\alias{forward}
\alias{pcount}
\alias{pmutate}
\alias{prename}
\title{Forwarders for overlapping functions in plyr and dplyr}
\usage{
pcount(df, vars = NULL, wt_var = NULL)

pmutate(.data, ...)

prename(x, replace, warn_missing = TRUE, warn_duplicated = TRUE)

dcount(x, ..., wt = NULL, sort = FALSE)

dmutate(.data, ...)

drename(.data, ...)
}
\arguments{
\item{df}{data frame to be processed}

\item{vars}{variables to count unique values of}

\item{wt_var}{optional variable to weight by - if this is non-NULL, count
will sum up the value of this variable for each combination of id
variables.}

\item{.data}{the data frame to transform}

\item{...}{named parameters giving definitions of new columns.}

\item{x}{named object to modify}

\item{replace}{named character vector, with new names as values, and
old names as names.}

\item{warn_missing}{print a message if any of the old names are
not actually present in \code{x}.}

\item{warn_duplicated}{print a message if any name appears more
  than once in \code{x} after the operation.
Note: x is not altered: To save the result, you need to copy the returned
  data into a variable.}

\item{wt}{(Optional) If not specified, will tally the number of rows.
If specified, will perform a "weighted" tally but summing over the
specified variable.}

\item{sort}{if \code{TRUE} will sort output in descending order of \code{n}}
}
\description{
Of the overlapping functions in \code{plyr} and \code{dplyr},
only three have a different interface and/or semantics:
\code{count}, \code{mutate} and \code{rename}.
These functions are forwarders to the corresponding functions
in the \code{plyr} and \code{dplyr} packages.
}
\seealso{
\code{plyr::\link[plyr]{count}}

\code{plyr::\link[plyr]{mutate}}

\code{plyr::\link[plyr]{rename}}

\code{dplyr::\link[dplyr]{count}}

\code{dplyr::\link[dplyr]{mutate}}

\code{dplyr::\link[dplyr]{rename}}
}

