% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap.R
\name{mutate}
\alias{mutate}
\title{Wrapped mutate function}
\usage{
mutate(.data, ...)
}
\arguments{
\item{.data}{the data frame to transform}

\item{...}{named parameters giving definitions of new columns.}
}
\description{
This active binding returns \code{get_mutate_wrapper()}.
Thus, the interface and the results of calling \code{mutate()}
from this package depend on the currently active compatibility layer.
When the package, is loaded the \code{\link{plyr_compat}} layer is active;
this documentation reflects the interface presented by that layer.
}
\seealso{
\code{\link{pdlyr_compat}}
}

