% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap.R
\name{rename}
\alias{rename}
\title{Wrapped rename function}
\usage{
rename(x, replace, warn_missing = TRUE, warn_duplicated = TRUE)
}
\arguments{
\item{x}{named object to modify}

\item{replace}{named character vector, with new names as values, and
old names as names.}

\item{warn_missing}{print a message if any of the old names are
not actually present in \code{x}.}

\item{warn_duplicated}{print a message if any name appears more
  than once in \code{x} after the operation.
Note: x is not altered: To save the result, you need to copy the returned
  data into a variable.}
}
\description{
This active binding returns \code{get_rename_wrapper()}.
Thus, the interface and the results of calling \code{rename()}
from this package depend on the currently active compatibility layer.
When the package, is loaded the \code{\link{plyr_compat}} layer is active;
this documentation reflects the interface presented by that layer.
}
\seealso{
\code{\link{pdlyr_compat}}
}

