% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compat.R
\docType{data}
\name{compat}
\alias{compat}
\alias{dplyr_compat}
\alias{plyr_compat}
\alias{plyr_warn_compat}
\title{Compatibility layers}
\format{\preformatted{List of 3
 $ mutate:function (.data, ...)  
 $ count :function (df, vars = NULL, wt_var = NULL)  
 $ rename:function (x, replace, warn_missing = TRUE)  
}}
\usage{
plyr_compat

plyr_warn_compat

dplyr_compat
}
\description{
TODO add documentation
}
\keyword{datasets}

