% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap.R
\name{pdlyr_compat}
\alias{get_pdlyr_compat}
\alias{pdlyr_compat}
\alias{set_pdlyr_compat}
\title{Active compatibility layer}
\usage{
get_pdlyr_compat()

set_pdlyr_compat(compat)
}
\arguments{
\item{x}{A compatibility layer, e.g., one described in \code{\link{compat}}}
}
\value{
The active or old compatibility layer
}
\description{
These functions set and return the active compatibility layer.
Functions from the active compatibility layer will be used when calling
\code{\link{count}}, \code{\link{mutate}} or \code{\link{rename}} from
this package.
}

